Move legacy test manager to 'Legacy' namespace
